Cambios Actualizados 10-12-2024

+++ BackEnd/CapaEntidad/Empleado.cs
@@ -7,7 +7,7 @@
         public string? cApellido { get; set; }
         public string? cPuesto { get; set; }
         public DateTime? dFechaContrato { get; set; }
-        public double? pSalario { get; set; }
+        public decimal? pSalario { get; set; }
 
     }
 }
